updated payment feature 0822
