+++ docker/mysql/data/binlog.index
@@ -11,3 +11,5 @@
 ./binlog.000011
 ./binlog.000012
 ./binlog.000013
+./binlog.000014
+./binlog.000015
