--- src/public/css/payment.css
@@ -1,9 +1,0 @@
-.payment-inner{
-    width: 40%;
-    margin:50px auto;
-}
-
-h2{
-    text-align: center;
-    margin-bottom: 50px;
-}
